rename 5d85cf0 parser for a bit
I will commit the old one back in next commit

This is to allow a github-based comparison

By the way, here is original commit:

https://github.com/williamsonrichard/records_in_contexts_draw_io_parser/blob/5d85cf05f47b3e14c3161a3ad2cd57f5fde67d09/draw_io_parser.py
